remove LeRobot dataset construction tooling
